requirements.txt
fastapi
uvicorn
twilio
python-dotenv
elevenlabs
starlette
python-multipart